+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 2021-05-31
+### Added
+* Image window UI.
+
+
 ## 2021-05-30
 ### Added
 * Ability to convert RT Image data and view this image inside the program.